question_answer_API/requirements.txt
Flask==1.0.2
Werkzeug==0.16.0
google-cloud-datastore==1.7.3
Click==7.0
Jinja2==2.10.1
MarkupSafe==1.1.1
WTForms==2.2.1
google-api-python-client
google-auth
google-auth-httplib2
requests
python-dateutil==2.8.1

